pyproject.toml
[build-system]
requires = [
 "setuptools>=68,<69",
  "wheel",
  "Cython>=0.29.28",
  "numpy>=1.18",
]
build-backend = "setuptools.build_meta"

[project]
name            = "lightning-stpp"
version         = "0.0.1"
license = { file = "LICENSE" }
description     = "Benchmarks for spatiotemporal point processes"
readme          = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch==2.6.0",
  "lightning==2.2.0",
  "ray[tune]==2.46.0",
  "numpy==2.3.1",
  "pandas==2.2.2",
  "torchdiffeq==0.2.5",
  "hydra-core==1.3.2",
  "omegaconf==2.3.0",
  "PyYAML==6.0.2",
  "mlflow==2.21.0",
  "tensorboard==2.18.0",
  "neural_stpp @ git+https://github.com/YahyaAalaila/neural_stpp.git@main#egg=neural_stpp",
  "deepstpp @ git+https://github.com/YahyaAalaila/DeepSTPP.git@master#egg=deepstpp",
  "smash @ git+https://github.com/YahyaAalaila/SMASH.git@main#egg=smash",
  "diffstpp @ git+https://github.com/YahyaAalaila/DiffSTPP.git@main#egg=diffstpp"
]
[tool.setuptools]              #  New
package-dir = {"" = "src"}     #  New

[tool.setuptools.packages.find]
where   = ["src"]
include = ["lightning_stpp", "lightning_stpp.*"]



